Merge pull-request #336 from FreeFeed:clbn/realtime-post-update to stable
Only update what is necessary on REALTIME_POST_UPDATE
This fixes the problem with post's likes and comments being updated (and
incorrectly cut/collapsed) after updating the post itself.

The same was fixed in 7c5140e for SAVE_EDITING_POST, but then the issue was
introduced again with the realtime updates.

Fixes #333.

[close #336]