Merge pull-request #378 from FreeFeed:clbn/fix-comment-edit to stable
Only clear editing comment state if necessary
Fixes T29, which is likely a regression after #335. That update successfully fixed #330 by clearing editText on saving, but it cleared it even when unnecessary, making the following editing broken and causing https://dev.freefeed.net/T29. So this update should fix it.

With this update, I tested adding, editing and multiple editing of comments both in feeds and on single post pages. Works on my machine.™

[close #378]