Properly clean auth cookie and localStorage key on log out
In previous version this code wrote 'undefined' as a string to localStorage. Also it didn't delete auth cookie (if it was present), so after reload the user remains logged in.

This version still displays 'User not found' just after log out from user's page (before the page reload), but it is unrelated to token storage.