Replace 'react-sticky' with custom component
It fixes #899.

The #899 bug was caused by the fact that the first and last comments were mounted separately from the others (which were included in StickyContainer). When the last comment becomes the penultimate one, it was re-mounted and reset the current editing status.

The new Sticky component does not require a container around the scrolling area and allows comments to be placed in a single list. This eliminates the need for re-mounting.