Add a general scroll compensating algorithm
After any scroll/resize/focus event sequence the page remembers ('pins') the focused input node or the node at the top of the screen. On every page update (controlling by MutationObserver and by special middleware) the position of pinned node is checking and compensating in case of change.

The realtime events are delaying until scroll or user interaction is finished.