Use CSS position: sticky in "Fold comments" control
The previous Sticky component polled/corrected the position of the elements on the 'scroll' and similar events. It worked pretty well on the desktop, but on mobile clients it led to big lags between scroll events and the control repositions.

The new implementation uses CSS "position: sticky". It still requires some JavaScript to detect the "stuck" status of the control and there is still some lag at this point. But the component itself is now tightly attached to the screen and does not move on mobile clients.