Rethink how RT events await user inactivity
In case of user activity or scroll, the old code collected realtime-actions in a special queue during the dispatching phase and re-dispatched them when the activity stopped. The activity check was performed on each dispatching, so if the activity was resumed at that moment then actions being transmitted endlessly to and from the queue. It was caused the browser to hang up.

The new code acts before dispatching and either dispatches the action immediately if there is no activity, or waits for the activity to stop for each action separately. After the activity is ended, the actions are dispatched without any further checks. This eliminates the possibility of an endless loop, since for each action the activity check is performed only once.