Detect updates of client code on the server
The client now can detect the updates of its code on the server and advice the user to reload the page. This feature is turned off by default.

When enabled, the client makes a HEAD request every 5 minutes (configurable) to the specific address ("/version.txt" is recommended) and looks for the specific response header ("Last-Modified" by default).

When the header value is changed, the information bar is shown to the user: “There’s a new update for FreeFeed available! _Refresh_the_page_ when you are ready.”