Let long links in comments be clickable on iOS
The problem was, that "comment-tail" span started on the same line as
comment text. Only the "space" character fit there and everything after
the dash was on second line and somehow, this span was blocking clicks.

I replaced spaces with non-breaking space here which forced comment-tail
to move to the next line.

The problem is gone now, but I'm not sure about exact reason of why it
helps.